Extract Python runtime resolution into its own module
