+++ src/python_session.rs
@@ -2,15 +2,12 @@ use std::collections::VecDeque;
 use std::ffi::{CStr, CString, c_char, c_int, c_long};
 #[cfg(target_family = "unix")]
 use std::os::unix::io::RawFd;
-use std::path::{Path, PathBuf};
-use std::process::{Command, Stdio};
+use std::path::Path;
 use std::ptr;
 #[cfg(target_family = "unix")]
 use std::sync::atomic::AtomicI32;
 use std::sync::atomic::{AtomicPtr, Ordering};
 use std::sync::{Arc, Condvar, Mutex, OnceLock};
-
-use serde::Deserialize;
 
 use crate::ipc;
 use crate::python_ffi::{GilGuard, ModuleMethod, PyObject, PyPtr, PyThreadState, PythonApi};
@@ -28,62 +25,8 @@ use windows_sys::Win32::System::Console::{GetStdHandle, STD_INPUT_HANDLE};
 #[cfg(windows)]
 use windows_sys::Win32::System::Pipes::PeekNamedPipe;
 
-pub const PYTHON_EXECUTABLE_ENV: &str = "MCP_REPL_PYTHON_EXECUTABLE";
 const MCP_REPL_PYTHON: &str = include_str!("../python/embedded.py");
 const PYTHON_EOF: c_int = 11;
-const PYTHON_PROGRAM: &str = "python3";
-const PYTHON_PROGRAM_FALLBACK: &str = "python";
-const PYTHON_CONFIG_SNIPPET: &str = r#"
-import json
-import sys
-import sysconfig
-
-def var(name):
-    value = sysconfig.get_config_var(name)
-    return "" if value is None else str(value)
-
-print(json.dumps({
-    "executable": sys.executable,
-    "base_executable": getattr(sys, "_base_executable", sys.executable),
-    "prefix": sys.prefix,
-    "base_prefix": sys.base_prefix,
-    "exec_prefix": sys.exec_prefix,
-    "base_exec_prefix": sys.base_exec_prefix,
-    "version": [sys.version_info[0], sys.version_info[1]],
-    "ldlibrary": var("LDLIBRARY"),
-    "instsoname": var("INSTSONAME"),
-    "libdir": var("LIBDIR"),
-    "libpl": var("LIBPL"),
-    "bindir": var("BINDIR"),
-    "pythonframeworkprefix": var("PYTHONFRAMEWORKPREFIX"),
-    "pythonframeworkinstalldir": var("PYTHONFRAMEWORKINSTALLDIR"),
-}))
-"#;
-
-#[derive(Debug)]
-struct PythonRuntimeConfig {
-    executable: PathBuf,
-    libpython: PathBuf,
-}
-
-#[derive(Debug, Deserialize)]
-struct PythonRuntimeProbe {
-    executable: String,
-    base_executable: String,
-    prefix: String,
-    base_prefix: String,
-    exec_prefix: String,
-    base_exec_prefix: String,
-    version: [u64; 2],
-    ldlibrary: String,
-    instsoname: String,
-    libdir: String,
-    libpl: String,
-    #[cfg(windows)]
-    bindir: String,
-    pythonframeworkprefix: String,
-    pythonframeworkinstalldir: String,
-}
 
 pub struct PythonSession {
     #[cfg(windows)]
@@ -649,7 +592,7 @@ fn run_session_on_current_thread(init: Arc<SessionInit>) -> Result<(), String> {
         return Err(message);
     }
 
-    let runtime_config = match resolve_python_runtime_config() {
+    let runtime_config = match crate::python_runtime::resolve_python_runtime_config() {
         Ok(runtime_config) => runtime_config,
         Err(err) => {
             init.mark_failed(err.clone());
@@ -818,313 +761,6 @@ fn set_stdio_unbuffered(file: *mut libc::FILE, fd: libc::c_int) -> Result<(), St
     } else {
         Err(format!("failed to configure worker fd {fd} as unbuffered"))
     }
-}
-
-fn find_dot_venv_pythons(start: &Path) -> Vec<PathBuf> {
-    let home = std::env::var_os("HOME").map(PathBuf::from);
-    // Search HOME itself, then stop. Do not ascend to HOME's parent.
-    let stop_at_home = home
-        .as_ref()
-        .filter(|home| start.starts_with(home.as_path()))
-        .cloned();
-    let mut dir = start.to_path_buf();
-    loop {
-        let mut candidates = Vec::new();
-        for candidate in [
-            dir.join(".venv").join("bin").join("python"),
-            dir.join(".venv").join("bin").join("python3"),
-        ] {
-            if candidate.is_file() {
-                candidates.push(candidate);
-            }
-        }
-        if !candidates.is_empty() {
-            return candidates;
-        }
-
-        if let Some(stop) = stop_at_home.as_ref()
-            && &dir == stop
-        {
-            break;
-        }
-
-        let Some(parent) = dir.parent() else {
-            break;
-        };
-        if parent == dir {
-            break;
-        }
-        dir = parent.to_path_buf();
-    }
-    Vec::new()
-}
-
-fn find_program_on_path(name: &str) -> Option<PathBuf> {
-    let path = std::env::var_os("PATH")?;
-    for dir in std::env::split_paths(&path) {
-        let candidate = dir.join(name);
-        if !candidate.is_file() {
-            continue;
-        }
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            if let Ok(meta) = std::fs::metadata(&candidate)
-                && meta.permissions().mode() & 0o111 != 0
-            {
-                return Some(candidate);
-            }
-        }
-
-        #[cfg(not(unix))]
-        {
-            return Some(candidate);
-        }
-    }
-    None
-}
-
-fn push_unique_path(paths: &mut Vec<PathBuf>, path: PathBuf) {
-    if !paths.iter().any(|existing| existing == &path) {
-        paths.push(path);
-    }
-}
-
-fn python_program_candidates() -> Vec<PathBuf> {
-    let mut candidates = Vec::new();
-    if let Ok(cwd) = std::env::current_dir() {
-        for venv_python in find_dot_venv_pythons(&cwd) {
-            push_unique_path(&mut candidates, venv_python);
-        }
-    }
-    push_unique_path(
-        &mut candidates,
-        find_program_on_path(PYTHON_PROGRAM).unwrap_or_else(|| PathBuf::from(PYTHON_PROGRAM)),
-    );
-    push_unique_path(
-        &mut candidates,
-        find_program_on_path(PYTHON_PROGRAM_FALLBACK)
-            .unwrap_or_else(|| PathBuf::from(PYTHON_PROGRAM_FALLBACK)),
-    );
-    candidates
-}
-
-fn query_python_runtime_config(executable: &Path) -> Result<PythonRuntimeConfig, String> {
-    let output = Command::new(executable)
-        .arg("-I")
-        .arg("-c")
-        .arg(PYTHON_CONFIG_SNIPPET)
-        .stdin(Stdio::null())
-        .output()
-        .map_err(|err| {
-            format!(
-                "failed to query Python runtime config from {}: {err}",
-                executable.display()
-            )
-        })?;
-    if !output.status.success() {
-        return Err(format!(
-            "failed to query Python runtime config from {}: {}",
-            executable.display(),
-            String::from_utf8_lossy(&output.stderr).trim()
-        ));
-    }
-    let probe: PythonRuntimeProbe = serde_json::from_slice(&output.stdout).map_err(|err| {
-        format!(
-            "failed to parse Python runtime config from {}: {err}",
-            executable.display()
-        )
-    })?;
-    let libpython = resolve_libpython_path(&probe).ok_or_else(|| {
-        format!(
-            "failed to locate a shared libpython for {}",
-            executable.display()
-        )
-    })?;
-    let executable = first_non_empty([probe.executable.as_str(), probe.base_executable.as_str()])
-        .map(PathBuf::from)
-        .unwrap_or_else(|| executable.to_path_buf());
-    Ok(PythonRuntimeConfig {
-        executable,
-        libpython,
-    })
-}
-
-#[cfg(test)]
-fn select_python_program(
-    mut candidates: Vec<PathBuf>,
-    mut starts: impl FnMut(&Path) -> bool,
-) -> PathBuf {
-    if candidates.is_empty() {
-        candidates.push(PathBuf::from(PYTHON_PROGRAM));
-    }
-    candidates
-        .iter()
-        .find(|candidate| starts(candidate))
-        .cloned()
-        .unwrap_or_else(|| candidates.remove(0))
-}
-
-fn select_python_runtime_config(
-    executable_override: Option<PathBuf>,
-    mut candidates: Vec<PathBuf>,
-    mut query: impl FnMut(&Path) -> Result<PythonRuntimeConfig, String>,
-) -> Result<PythonRuntimeConfig, String> {
-    if let Some(executable) = executable_override {
-        return query(&executable);
-    }
-
-    if candidates.is_empty() {
-        candidates.push(PathBuf::from(PYTHON_PROGRAM));
-    }
-
-    let mut errors = Vec::new();
-    for candidate in candidates {
-        match query(&candidate) {
-            Ok(config) => return Ok(config),
-            Err(err) => errors.push(format!("{}: {err}", candidate.display())),
-        }
-    }
-
-    Err(format!(
-        "failed to query Python runtime config from candidate interpreters: {}",
-        errors.join("; ")
-    ))
-}
-
-fn resolve_python_runtime_config() -> Result<PythonRuntimeConfig, String> {
-    select_python_runtime_config(
-        std::env::var_os(PYTHON_EXECUTABLE_ENV).map(PathBuf::from),
-        python_program_candidates(),
-        query_python_runtime_config,
-    )
-}
-
-fn resolve_libpython_path(probe: &PythonRuntimeProbe) -> Option<PathBuf> {
-    let mut candidates = Vec::new();
-    push_python_library_candidates(&mut candidates, probe, &probe.ldlibrary);
-    if probe.instsoname != probe.ldlibrary {
-        push_python_library_candidates(&mut candidates, probe, &probe.instsoname);
-    }
-    push_windows_python_library_candidates(&mut candidates, probe);
-
-    let version = format!("{}.{}", probe.version[0], probe.version[1]);
-    for root in [
-        probe.base_exec_prefix.as_str(),
-        probe.exec_prefix.as_str(),
-        probe.base_prefix.as_str(),
-        probe.prefix.as_str(),
-    ] {
-        if root.is_empty() {
-            continue;
-        }
-        candidates.push(
-            Path::new(root)
-                .join("lib")
-                .join(format!("libpython{version}.so")),
-        );
-        candidates.push(
-            Path::new(root)
-                .join("lib")
-                .join(format!("libpython{version}.dylib")),
-        );
-        candidates.push(Path::new(root).join("Python"));
-    }
-
-    candidates
-        .into_iter()
-        .find(|candidate| is_loadable_libpython_candidate(candidate))
-}
-
-fn is_loadable_libpython_candidate(candidate: &Path) -> bool {
-    if !candidate.is_file() {
-        return false;
-    }
-    !candidate
-        .extension()
-        .and_then(|extension| extension.to_str())
-        .is_some_and(|extension| {
-            extension.eq_ignore_ascii_case("a") || extension.eq_ignore_ascii_case("lib")
-        })
-}
-
-#[cfg(windows)]
-fn push_windows_python_library_candidates(
-    candidates: &mut Vec<PathBuf>,
-    probe: &PythonRuntimeProbe,
-) {
-    let compact_version = format!("{}{}", probe.version[0], probe.version[1]);
-    let library_names = [
-        format!("python{compact_version}.dll"),
-        "python3.dll".to_string(),
-    ];
-
-    for root in [
-        Path::new(probe.executable.as_str()).parent(),
-        Path::new(probe.base_executable.as_str()).parent(),
-        non_empty(&probe.bindir).map(Path::new),
-        non_empty(&probe.base_exec_prefix).map(Path::new),
-        non_empty(&probe.exec_prefix).map(Path::new),
-        non_empty(&probe.base_prefix).map(Path::new),
-        non_empty(&probe.prefix).map(Path::new),
-    ]
-    .into_iter()
-    .flatten()
-    {
-        for library in &library_names {
-            candidates.push(root.join(library));
-        }
-    }
-}
-
-#[cfg(not(windows))]
-fn push_windows_python_library_candidates(
-    _candidates: &mut Vec<PathBuf>,
-    _probe: &PythonRuntimeProbe,
-) {
-}
-
-fn push_python_library_candidates(
-    candidates: &mut Vec<PathBuf>,
-    probe: &PythonRuntimeProbe,
-    library: &str,
-) {
-    let Some(library) = non_empty(library) else {
-        return;
-    };
-    let path = Path::new(library);
-    if path.is_absolute() {
-        candidates.push(path.to_path_buf());
-    }
-    for executable in [probe.executable.as_str(), probe.base_executable.as_str()] {
-        let Some(executable) = non_empty(executable) else {
-            continue;
-        };
-        let Some(parent) = Path::new(executable).parent() else {
-            continue;
-        };
-        candidates.push(parent.join(library));
-    }
-    for root in [
-        probe.libdir.as_str(),
-        probe.libpl.as_str(),
-        probe.pythonframeworkprefix.as_str(),
-        probe.pythonframeworkinstalldir.as_str(),
-    ] {
-        if let Some(root) = non_empty(root) {
-            candidates.push(Path::new(root).join(library));
-        }
-    }
-}
-
-fn first_non_empty<'a>(values: impl IntoIterator<Item = &'a str>) -> Option<&'a str> {
-    values.into_iter().find_map(non_empty)
-}
-
-fn non_empty(value: &str) -> Option<&str> {
-    let value = value.trim();
-    if value.is_empty() { None } else { Some(value) }
 }
 
 fn initialize_python(
@@ -2943,39 +2579,6 @@ static PYTHON_RUNTIME_STDIN_FD: AtomicI32 = AtomicI32::new(-1);
 mod tests {
     use super::*;
 
-    fn runtime_config_for(path: &str) -> PythonRuntimeConfig {
-        PythonRuntimeConfig {
-            executable: PathBuf::from(path),
-            libpython: PathBuf::from("python.dll"),
-        }
-    }
-
-    fn runtime_probe_for_libpython(
-        executable: &Path,
-        version: [u64; 2],
-        ldlibrary: &Path,
-        libdir: &Path,
-        prefix: &Path,
-    ) -> PythonRuntimeProbe {
-        PythonRuntimeProbe {
-            executable: executable.to_string_lossy().into_owned(),
-            base_executable: executable.to_string_lossy().into_owned(),
-            prefix: prefix.to_string_lossy().into_owned(),
-            base_prefix: prefix.to_string_lossy().into_owned(),
-            exec_prefix: prefix.to_string_lossy().into_owned(),
-            base_exec_prefix: prefix.to_string_lossy().into_owned(),
-            version,
-            ldlibrary: ldlibrary.to_string_lossy().into_owned(),
-            instsoname: ldlibrary.to_string_lossy().into_owned(),
-            libdir: libdir.to_string_lossy().into_owned(),
-            libpl: libdir.to_string_lossy().into_owned(),
-            #[cfg(windows)]
-            bindir: String::new(),
-            pythonframeworkprefix: String::new(),
-            pythonframeworkinstalldir: String::new(),
-        }
-    }
-
     #[cfg(target_family = "unix")]
     fn active_request_for_prompt_wait(
         line_count: usize,
@@ -3035,110 +2638,5 @@ mod tests {
             &active,
             Some(PythonReadlineState::Primary)
         ));
-    }
-
-    #[test]
-    fn python_program_selection_falls_back_after_broken_python3_candidate() {
-        let selected = select_python_program(
-            vec![PathBuf::from("python3"), PathBuf::from("python")],
-            |candidate| candidate == Path::new("python"),
-        );
-
-        assert_eq!(selected, PathBuf::from("python"));
-    }
-
-    #[test]
-    fn python_runtime_config_falls_back_after_broken_python3_candidate() {
-        let mut attempts = Vec::new();
-
-        let config = select_python_runtime_config(
-            None,
-            vec![PathBuf::from("python3"), PathBuf::from("python")],
-            |candidate| {
-                attempts.push(candidate.to_path_buf());
-                if candidate == Path::new("python3") {
-                    Err("store alias is not a usable interpreter".to_string())
-                } else {
-                    Ok(runtime_config_for("python"))
-                }
-            },
-        )
-        .expect("python fallback should be used after python3 fails");
-
-        assert_eq!(
-            attempts,
-            vec![PathBuf::from("python3"), PathBuf::from("python")]
-        );
-        assert_eq!(config.executable, PathBuf::from("python"));
-    }
-
-    #[test]
-    fn python_runtime_config_env_override_does_not_fallback() {
-        let mut attempts = Vec::new();
-
-        let err = select_python_runtime_config(
-            Some(PathBuf::from("custom-python")),
-            vec![PathBuf::from("python3"), PathBuf::from("python")],
-            |candidate| {
-                attempts.push(candidate.to_path_buf());
-                Err(format!("{} is not usable", candidate.display()))
-            },
-        )
-        .expect_err("explicit Python override should not fall back");
-
-        assert_eq!(attempts, vec![PathBuf::from("custom-python")]);
-        assert!(err.contains("custom-python is not usable"));
-    }
-
-    #[test]
-    fn resolve_libpython_path_skips_static_archive_candidate() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let bin = temp.path().join("bin");
-        let lib = temp.path().join("lib");
-        std::fs::create_dir_all(&bin).expect("bin dir");
-        std::fs::create_dir_all(&lib).expect("lib dir");
-        let executable = bin.join("python3");
-        let archive = lib.join("libpython3.11.a");
-        let shared = lib.join("libpython3.11.so");
-        std::fs::write(&executable, "").expect("python placeholder");
-        std::fs::write(&archive, "!<arch>\n").expect("archive placeholder");
-        std::fs::write(&shared, "").expect("shared placeholder");
-
-        let probe = runtime_probe_for_libpython(&executable, [3, 11], &archive, &lib, temp.path());
-
-        assert_eq!(resolve_libpython_path(&probe), Some(shared));
-    }
-
-    #[cfg(windows)]
-    #[test]
-    fn resolve_libpython_path_finds_windows_dll_next_to_executable() {
-        fn runtime_probe_for(executable: &str) -> PythonRuntimeProbe {
-            PythonRuntimeProbe {
-                executable: executable.to_string(),
-                base_executable: executable.to_string(),
-                prefix: String::new(),
-                base_prefix: String::new(),
-                exec_prefix: String::new(),
-                base_exec_prefix: String::new(),
-                version: [3, 11],
-                ldlibrary: String::new(),
-                instsoname: String::new(),
-                libdir: String::new(),
-                libpl: String::new(),
-                bindir: String::new(),
-                pythonframeworkprefix: String::new(),
-                pythonframeworkinstalldir: String::new(),
-            }
-        }
-
-        let temp = tempfile::tempdir().expect("tempdir");
-        let python = temp.path().join("python.exe");
-        let dll = temp.path().join("python311.dll");
-        std::fs::write(&python, "").expect("python placeholder");
-        std::fs::write(&dll, "").expect("dll placeholder");
-
-        let probe = runtime_probe_for(&python.to_string_lossy());
-
-        assert_eq!(resolve_libpython_path(&probe), Some(dll));
     }
 }
